requirements.txt
langchain==0.3.9
langchain-community==0.3.1
pypdf==4.2.0
python-dotenv
tiktoken==0.7.0
langchain-experimental==0.3.3
streamlit==1.35.0
faiss-cpu==1.10.0
boto3==1.35.7
openai==1.42.0
httpx==0.27.0
tavily-python==0.5.1